requirements.txt
langchain
openai
uvicorn
python-dotenv
chromadb
unstructured[pdf]
sentence-transformers
